refactor(fsd): api client と schema.gen を shared/api へ移設 (#356)
A2 Lane2 shared foundation の api segment(Tier1・機械的移設)。client.ts と
生成型 schema.gen.ts を shared/api へ。endpoints.ts は entities への分解
(後続 PR)までは api/ に残置し、client 参照だけ shared を指すよう更新。

- src/api/client.ts(+test) → src/shared/api/client.ts
- src/api/schema.gen.ts → src/shared/api/schema.gen.ts
- @/api/client の import 8箇所 + endpoints の相対 './client' を
  @/shared/api/client に書換(残存0)。
- codegen インフラのパス3点を追随: package.json の codegen 出力先・
  scripts/check-codegen.mjs の COMMITTED・knip.json の ignore。

挙動不変(move + import/パス書換のみ・ロジック diff ゼロ)。
**npm run check 全体緑**(codegen:check 込み・#361 の教訓)・62 tests 緑。
