+++ frontend/src/shared/api/schema.gen.ts
